こみっと

+++ guessing_game/src/main.rs
@@ -16,7 +16,7 @@ fn main() {
             .read_line(&mut guess)
             .expect("読み込みに失敗");
     
-        let guess: u32 = match guess.trim().parse() {
+        let guess: u8 = match guess.trim().parse() {
             Ok(num) => num,
             Err(_) => continue,
         };
